fix: idiot formatting
Whether that be by fixing map-printing[1] or factorials[2].

[1]: map-printing
By map printing, I -- of course -- mean the printing of the map by using
the printMap() function. What was -- arguably -- "fixed" was newlines
and indents in the map printing, e.g if the map is equal to 1, "A"; 2,
"B"; 3, "C"; 4, "D" and 5, "E"", this would be the old printMap()
functions output:

{1:"A" 2:"B" 3:"C" 4:"D" 5:"E"}

and this is -- though uhh, sadly untested -- the new one:

{
    1 : "A"
    2 : "B"
    3 : "C"
    4 : "D"
    5 : "E"
}

[2]: factorials
What we refer to as the "factorial" is -- in reality -- annotated as
f(). The "idiot" part of "idiot formatting" comes in here. We need to
check if "n" (or: @param 1) is equal to one. I -- because I thought I
was clever -- wrote this:

    (n) ? a : b

My belief was that this would make

    if (n == 1) {
        a
    } else {
        b
    }

And I still think it does, don't get me wrong! The only problem is that
this is *very* ugly. I could easily fix this though[3].

[3]: Easily fixing this though
(n is 1) ? a : b
-> (n == 1) ? a : b
-> if (n == 1) {
      a
   } else {
      b
   }